Tests/LibWeb/Text/expected/css/font-face-activation-after-stylesheet-media-change.txt

Initial face count: 0
Matching face count: 1
Matching face status: loaded
Final face count: 0
